Add permissions checks

+++ projects/views.py
@@ -1,22 +1,25 @@
-from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework import viewsets
 from rest_framework import status
 from .models import Project, Task
 from .serializers import ProjectSerializer, TaskSerializer
+from .permissons import IsProjectMember
 from django.shortcuts import get_object_or_404
 
 
 class TaskViewSet(viewsets.ModelViewSet):
     queryset = Task.objects.all()
     serializer_class = TaskSerializer
-    permission_classes = [IsAuthenticatedOrReadOnly]
+    permission_classes = [IsAuthenticated]
 
 
 class ProjectListCreateApiView(APIView):
+    permission_classes = [IsAuthenticated]
+    
     def get(self, request):
-        projects = Project.objects.filter(members=request.user)
+        projects = Project.objects.all()
         serializer = ProjectSerializer(projects, many=True)
         return Response(serializer.data)
 
@@ -30,16 +33,20 @@ class ProjectListCreateApiView(APIView):
 
 
 class ProjectDetailApiView(APIView):
+    permission_classes = [IsAuthenticated, IsProjectMember]
+    
     def get_object(self, pk):
         return get_object_or_404(Project, pk=pk)
 
-    def get(self, pk):
+    def get(self, request, pk):
         project = self.get_object(pk)
+        self.check_object_permissions(request, project)
         serializer = ProjectSerializer(project)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def put(self, request, pk):
         project = self.get_object(pk)
+        self.check_object_permissions(request, project)
         serializer = ProjectSerializer(project, data=request.data)
 
         if serializer.is_valid():
@@ -47,7 +54,8 @@ class ProjectDetailApiView(APIView):
             return Response(serializer.data, status=status.HTTP_200_OK)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, pk):
+    def delete(self, request, pk):
         project = self.get_object(pk)
+        self.check_object_permissions(request, project)
         project.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
